Move app.py to noauth.py
